rotation check function
Hi Ania.  I really enjoyed your tutorial on youtube and found it very helpful.  I think I have a fix for rotation sometimes causing pieces to cross the screen.  It could probably be more elegant but it seems to work (I included left/right edge checks as functions because it was easier for me that way).